requirements.txt
ailment==8.18.10.5
aiohttp==1.0.5
amazon-dash==1.1.1
ana==0.5
angr==8.18.10.5
archinfo==8.18.10.5
asn1crypto==0.24.0
async-timeout==3.0.1
bcrypt==3.1.4
beautifulsoup4==4.6.3
bitstring==3.1.5
cachetools==2.1.0
capstone==3.0.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
claripy==8.18.10.5
cle==8.18.10.5
click==6.7
click-default-group==1.2
cooldict==1.3
crypto==1.4.1
cryptography==2.3.1
cycler==0.10.0
decorator==4.0.11
discord.py==0.16.12
dnspython==1.15.0
dpkt==1.9.1
Flask==1.0.2
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
gmpy==1.17
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.14
gym==0.10.5
hashpumpy==1.2
homebrew-pypi-poet==0.10.0
httplib2==0.12.0
idalink==0.12
idna==2.7
iexfinance==0.3.4
imageio==2.1.2
impacket==0.9.17
InstagramAPI==1.0.2
intervaltree==2.1.0
ItsDangerous==1.0.0
Jinja2==2.10
jsonschema==2.6.0
kiwisolver==1.0.1
ldap3==2.5.1
ldapdomaindump==0.9.0
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
meson==0.49.1
moviepy==0.2.3.2
msgpack==0.5.6
mulpyplexer==0.8
multidict==4.4.2
Naked==0.1.31
neovim==0.2.6
networkx==2.2
numpy==1.14.5
oauthlib==3.0.1
packaging==18.0
pandas==0.23.3
paramiko==2.4.1
parse==1.9.0
pefile==2018.8.8
Pillow==5.3.0
pluggy==0.7.1
plumbum==1.6.7
progressbar==2.5
protobuf==3.6.1
psutil==5.4.7
psycopg2==2.7.6.1
pwntools==3.12.1
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycparser==2.18
pycrypto==2.6.1
pyelftools==0.24
pyglet==1.3.2
Pygments==2.2.0
PyNaCl==1.2.1
pyOpenSSL==18.0.0
pypandoc==1.4
pyparsing==2.2.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
pyvex==8.18.10.5
PyYAML==3.13
requests==2.11.1
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
ROPGadget==5.4
rpyc==4.0.2
rsa==4.0
rsatool==1.0
scapy==2.4.0
scikit-learn==0.19.2
scipy==1.1.0
shellescape==3.4.1
six==1.11.0
sklearn==0.0
smmap2==2.0.5
sortedcontainers==1.5.10
tox==3.2.1
tqdm==4.11.2
unicorn==1.0.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virustotal-api==1.1.10
websockets==3.4
Werkzeug==0.14.1
yapf==0.25.0
z3-solver==4.5.1.0.post2
zybook==1.0
